Use a trait for traces, so that different backends can return different traces.
